Folder refactor for better separation
